requirements.txt
tensorflow==2.12.0
pandas
dvc
mlflow
notebook
numpy
matplotlib
seaborn
python-box
pyYAML
tqdm
ensure
joblib
types-pyYAML
scipy
Flask
Flask-Cors
gdown
-e .